// client-side/src/app/portals/Track-Force/models/status-info.model.ts
export class StatusInfo {
    name: string;
    trainingMapped: number;
	trainingUnmapped: number;
    reservedMapped: number;
	openUnmapped: number;
	selectedMapped: number;
	selectedUnmapped: number;
	confirmedMapped: number;
	confirmedUnmapped: number;
	deployedMapped: number;
	deployedUnmapped: number;
}